fashion_search_bot/requirements.txt
# Core dependencies
fastapi==0.111.1
uvicorn==0.23.2
jinja2==3.1.6
python-dotenv==1.0.1
requests==2.32.5

# Image processing & numerical operations
Pillow==11.3.0
numpy==2.3.3
torch  # Install the latest compatible version for your system
transformers  # For BLIP or other transformer models

# AI / ML utilities
ultralytics  # YOLO if needed
serpapi  # For scraping search results (SerpApi Python SDK)
